Add files via upload
A social media share button repository using HTML and CSS would likely contain the following files:

index.html - This is the main HTML file that contains the markup for the share button. It would include the necessary HTML tags such as <head>, <body>, <div>, and <button> to create the structure of the button.

style.css - This is the CSS file that contains the styles for the share button. It would include the necessary CSS properties such as font-size, color, background-color, padding, and border-radius to style the button to the desired design.